Read the whole output buffer from child process
Currently, we will read only a part of the ouput.
TO be specific only the last line after the last
stdin. Also we will save every current test in
the file in JSON format with key -> current test
number and value -> stdout buffer. If we use the
script from this commit it will NOT WORK, because
we override the file every time. This will be
fixed in the next commits.